% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbfc.R
\name{cbf.c}
\alias{cbf.c}
\title{CBF lazy bloody list}
\usage{
cbf.c(listargs, split = " ", fixed = T)
}
\arguments{
\item{listargs}{a string containing the list items}

\item{split}{character to split by. If FALSE, a character vector (or object which can be coerced to such) containing regular expression(s) to use for splitting. If empty matches occur, in particular if split has length 0, x is split into single characters. If split has length greater than 1, it is re-cycled along x.}

\item{fixed}{passed to strsplit. logical. If TRUE match split exactly, otherwise use regular expressions. Has priority over perl.}
}
\value{

}
\description{
CBF lazy bloody list
}
