% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbfc.R
\name{cbf.coltype_numeric}
\alias{cbf.coltype_numeric}
\title{CBF Mutate Type to numeric}
\usage{
cbf.coltype_numeric(.x, .cbfcols, ...)
}
\arguments{
\item{.x}{A dataframe (tibble) with column to be mutated}

\item{.cbfcols}{A [cbf.c()] or string of column name(s) to mutate}
}
\description{
CBF Mutate Type to numeric
}
\seealso{
Other coltype: 
\code{\link{cbf.coltype_char}()},
\code{\link{cbf.coltype_factor}()},
\code{\link{cbf.coltype_int}()},
\code{\link{cbf.coltype}()}
}
\concept{coltype}
