% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbf_stats.R
\name{cbf_statmode}
\alias{cbf_statmode}
\title{Returns mode of a variable}
\source{
\url{https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}
}
\usage{
cbf_statmode(x)
}
\arguments{
\item{x}{vector.}
}
\value{
Most common value.
}
\description{
This function takes a vector and returns the statistical mode, i.e. the most
 common value the vector takes.
}
\examples{
stat_mode(ggplot2::mpg$manufacturer)
stat_mode(ggplot2::mpg$model)

stat_mode(rpois(100, 1))
}
